CDAP-14905 move program runtime service into spi module
Moving most of the ProgramRuntimeService and associated interfaces
out of app-fabric and into master-spi in order to prepare for a
kubernetes implementation.

Besides moving classes around, a change was made to the
ProgramController to return Futures instead of ListenableFutures
in order to remove the dependency on guava. This required changing
ProgramLifecycleService.issueStop() from an async call to a
synchronous call. Everything that was calling it was immediately
getting the future, so everything still behaves as it did before.